requirements.txt
elevenlabs==0.2.24
python-dotenv==1.0.0
pillow==10.1.0
pygame==2.5.2 